ReverseSSH: add support to kill ssh server when ssh client exits ( OFW-3704 )
SSH server should be killed after SSH client process exits

[Issue]
as per CGW-19, ssh server must be terminated after ssh client exits for any reason.

[Solution]
Run ssh client in a subshell in foreground and kill ssh server after
ssh client process exits for any reason.

[Testing Performed]
- Enable RSSH with wrong parameters and let it fail.
- check ps output and verify that the SSH server is not running.

Upstream Candidate: Maybe, Modification is as per standard.

+++ startTunnel.sh
@@ -31,9 +31,31 @@ fi
 SSH_CLIENT="/usr/bin/dbclient"
 SSH_DAEMON="/usr/sbin/dropbear"
 WAN_INTERFACE=$(getWanInterfaceName)
+STARTUNNELSH=$0
+
 usage()
 {
   echo "USAGE:   startTunnel.sh {start|stop} {args}"
+}
+
+startSshClientandDaemon()
+{
+  fd=205
+  lockFile="/var/lock/$(basename $0)"
+  eval exec "$fd"'>"$lockFile"'
+
+  flock "$fd"
+  (
+    # Start SSH daemon on demand
+    start-stop-daemon -S -x $SSH_DAEMON -m -b -p /var/tmp/rsshd.pid -- -B -F
+
+    if [ -z "$passwd" ]; then
+      start-stop-daemon -S -x $SSH_CLIENT -m -p /var/tmp/rssh.pid -- $args
+    else
+      DROPBEAR_PASSWORD="$passwd" start-stop-daemon -S -x $SSH_CLIENT -m -p /var/tmp/rssh.pid -- $args
+    fi
+    $STARTUNNELSH stop
+  )
 }
 
 if [ -f /lib/rdk/utils.sh  ]; then
@@ -63,10 +85,11 @@ case $oper in
              ;;
            start)
             if [ "$FIRMWARE_TYPE" = "OFW" ]; then
+                # Kill and ssh daemon not started by us
                 killall `basename $SSH_DAEMON`
 
-                # Start SSH daemon on demand
-                start-stop-daemon -S -x $SSH_DAEMON -m -b -p /var/tmp/rsshd.pid -- -B -F
+                #Cleanup previous session if any
+                $STARTUNNELSH stop
 
                 args=`echo $*`
                 if [[ "$*" =~ "passwd=" ]]; then
@@ -75,12 +98,7 @@ case $oper in
                     args="$(awk '{$1 = ""; print $0;}' <<<$args)" # Update SSH args to exclude login password
                 fi
 
-                if [ -z "$passwd" ]; then
-                    start-stop-daemon -S -x $SSH_CLIENT -m -b -p /var/tmp/rssh.pid -- $args
-                else
-                    DROPBEAR_PASSWORD="$passwd" start-stop-daemon -S -x $SSH_CLIENT -m -b -p /var/tmp/rssh.pid -- $args
-                fi
-                sleep 10
+                startSshClientandDaemon &
                 exit 1
             fi
          WAN_INTERFACE=$(getWanInterfaceName)
@@ -188,10 +206,14 @@ case $oper in
            stop)
              if [ "$FIRMWARE_TYPE" = "OFW" ]; then
                  # use start-stop-daemon to kill rssh and sshd
-                 start-stop-daemon -K -p /var/tmp/rssh.pid
-                 rm /var/tmp/rssh.pid
-                 start-stop-daemon -K -p /var/tmp/rsshd.pid
-                 rm /var/tmp/rsshd.pid
+                 if [ -f /var/tmp/rssh.pid ];then
+                     start-stop-daemon -K -p /var/tmp/rssh.pid
+                     rm /var/tmp/rssh.pid
+                 fi
+                 if [ -f /var/tmp/rsshd.pid ];then
+                     start-stop-daemon -K -p /var/tmp/rsshd.pid
+                     rm /var/tmp/rsshd.pid
+                 fi
                  exit 1
              fi
              cat /var/tmp/rssh.pid |xargs kill -9
